vol002-opencv/requirements.txt
PyYAML==3.13
RPi.GPIO==0.6.5
ffmpy==0.2.2
google-api-python-client==1.7.4
oauth2client==4.1.3
opencv-python==3.4.3.18
picamera==1.13
pillow==5.3.0
slackclient==1.3.0
